[Raagul] Update : Updated Delete Function in the Application

+++ src/main/java/org/bridgelabz/reactivegreetingapplication/service/Services.java
@@ -37,7 +37,16 @@ public class Services implements ServiceInterface{
     }
 
     @Override
-    public Mono<Void> deleteGreetings(int id) {
-        return repository.deleteById(id);
+    public Mono<Boolean> deleteGreetings(int id) {
+        return repository.existsById(id)
+                .flatMap(exists->{
+                    if(exists){
+                        return repository.deleteById(id)
+                                .thenReturn(true);
+                    }
+                    else {
+                        return Mono.just(false);
+                    }
+                });
     }
 }
